broke test validation out into independent files
