+++ tests/EN_range_check_validation.py
@@ -1,0 +1,37 @@
+import qctests.EN_range_check
+
+import util.testingProfile
+import numpy
+
+##### EN_range_check ---------------------------------------------------
+
+def test_EN_range_check_temperature():
+    '''
+    Make sure EN_range_check is flagging temperature excursions
+    '''
+
+    # should fail despite rounding
+    p = util.testingProfile.fakeProfile([-4.00000001], [100]) 
+    qc = qctests.EN_range_check.test(p)
+    truth = numpy.zeros(1, dtype=bool)
+    truth[0] = True
+    assert numpy.array_equal(qc, truth), 'failed to flag temperature slightly colder than -4 C'
+
+    # -4 OK
+    p = util.testingProfile.fakeProfile([-4], [100]) 
+    qc = qctests.EN_range_check.test(p)
+    truth = numpy.zeros(1, dtype=bool)
+    assert numpy.array_equal(qc, truth), 'incorrectly flagging -4 C'
+
+    # 40 OK
+    p = util.testingProfile.fakeProfile([40], [100]) 
+    qc = qctests.EN_range_check.test(p)
+    truth = numpy.zeros(1, dtype=bool)
+    assert numpy.array_equal(qc, truth), 'incorrectly flagging 40 C'
+
+    # should fail despite rounding
+    p = util.testingProfile.fakeProfile([40.0000001], [100]) 
+    qc = qctests.EN_range_check.test(p)
+    truth = numpy.zeros(1, dtype=bool)
+    truth[0] = True
+    assert numpy.array_equal(qc, truth), 'failed to flag temperature slightly warmer than 40 C'
